requirements.txt
fastapi==0.115.12
python-multipart==0.0.22
uvicorn==0.30.6
sqlalchemy==2.0.39
alembic==1.14.1
pydantic-settings==2.8.1
PyJWT==2.10.1
psycopg[binary]==3.2.13
pytest==8.3.5
httpx==0.28.1
requests==2.32.3
ruff==0.11.5
black==25.1.0